+++ hash_map.h
@@ -17,7 +17,7 @@
 struct hash_map *hash_map_create(void);
 void hash_map_destroy(struct hash_map *map);
 void hash_map_insert(struct hash_map *map, char *key, void *val);
-int hash_map_at(struct hash_map *map, char *key, void **val);
+int64_t hash_map_at(struct hash_map *map, char *key, void **val);
 uint64_t hash_map_length(struct hash_map *map);
 uint64_t hash_map_capacity(struct hash_map *map);
 struct hash_map_iter *hash_map_iter_create(struct hash_map *map);
